Add graph.py for visualization generation and enhanced filtering on analysis tab
